% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aapply}
\alias{aapply}
\title{aapply}
\usage{
aapply(SE, fx, assay_type = 1)
}
\arguments{
\item{SE}{a SummarizedExperiment object}

\item{fx}{any function}

\item{assay_type}{a name of an assay}
}
\value{
the same SE object with updated nested dataframe
}
\description{
Works like sapply but on each nested dataframe of the assay of an SE
}
