% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{add_CellLine_annotation}
\alias{add_CellLine_annotation}
\title{add_CellLine_annotation}
\usage{
add_CellLine_annotation(df_metadata, fill = NULL)
}
\arguments{
\item{df_metadata}{a data.frame with metadata}

\item{fill}{a string indicating how unknown cell lines should be filled in the DB}
}
\value{
a data.frame with metadata with annotated cell lines
}
\description{
add cellline annotation to a data.frame with metadata
}
\details{
The logic of adding celline annotation for df_metadata is based on the function
get_cell_lines from the gDRwrapper
we added additional parameter 'fill' that allows to fill the DB with clid info for these cell lines
that are not present in the DB.
Other fields are set as "UNKNOWN". If the fill is set as FALSE we add unknown cell lines
only to the tibble.
This approach will be corrected once we will implement final solution for adding cell lines.
}
