% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{add_codrug_group_SE}
\alias{add_codrug_group_SE}
\title{Add co-drug group to SE}
\usage{
add_codrug_group_SE(se)
}
\arguments{
\item{se}{a BumpyMatrix SE with drug response data}
}
\value{
SE object
}
\description{
Add co-drug group to SE
}
