% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{add_Drug_annotation}
\alias{add_Drug_annotation}
\title{add_Drug_annotation}
\usage{
add_Drug_annotation(df_metadata, fname = "drugs.csv", fill = "unknown")
}
\arguments{
\item{df_metadata}{a data.frame with metadata}

\item{fname}{a string with file name with annotation}

\item{fill}{a string indicating how unknown cell lines should be filled in the DB}
}
\value{
a data.frame with metadata with annotated drugs
}
\description{
add drug annotation to a data.frame with metadata
}
\details{
The logic of adding drug annotation for df_metadata based on the annotation file stored
in gDRtestData.
}
