% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_intermediate_data}
\alias{add_intermediate_data}
\title{add intermediate data (qs2 files) for given MAE}
\usage{
add_intermediate_data(mae, data_dir, steps = get_pipeline_steps())
}
\arguments{
\item{mae}{\code{MultiAssayExperiment} with dose-response data}

\item{data_dir}{output directory}

\item{steps}{character vector with pipeline steps for which
intermediate data should be saved}
}
\value{
\code{NULL}
}
\description{
add intermediate data (qs2 files) for given MAE
}
\keyword{internal}
