% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{assert_cell_line_annotation}
\alias{assert_cell_line_annotation}
\title{Assert cell line annotation}
\usage{
assert_cell_line_annotation(cell_line_annotation)
}
\arguments{
\item{cell_line_annotation}{data.table with cell line annotations}
}
\description{
Validates that the cell line annotation data.table has the required columns.
}
\keyword{internal}
