% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{assert_drug_annotation}
\alias{assert_drug_annotation}
\title{Assert drug annotation}
\usage{
assert_drug_annotation(drug_annotation)
}
\arguments{
\item{drug_annotation}{data.table with drug annotations}
}
\description{
Validates that the drug annotation data.table has the required columns.
}
\keyword{internal}
