% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_combo_codilution.R
\name{calculate_combo_codilution}
\alias{calculate_combo_codilution}
\title{Calculate combo co-dilution}
\usage{
calculate_combo_codilution(se)
}
\arguments{
\item{se}{a BumpyMatrix SE with drug response data}
}
\value{

}
\description{
Calculate combo co-dilution
}
