% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{calculate_combo_matrix}
\alias{calculate_combo_matrix}
\title{Calculate combo matrix}
\usage{
calculate_combo_matrix(
  se,
  nested_identifiers,
  conc_margin = 10^0.5,
  log2_pos_offset = log10(3)/2,
  norm_types = c("RelativeViability", "GRvalue"),
  averaged_assay = "Averaged"
)
}
\arguments{
\item{se}{a BumpyMatrix SE with drug response data}

\item{nested_identifiers}{series identifiers}

\item{conc_margin}{concentration margin}

\item{log2_pos_offset}{logarithm offset}

\item{norm_types}{types of normalization}

\item{averaged_assay}{name of averaged assay}
}
\value{
list with results
}
\description{
Calculate combo matrix
}
