% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanup_metadata}
\alias{cleanup_metadata}
\title{cleanup_metadata}
\usage{
cleanup_metadata(
  df_metadata,
  cell_line_annotation = NULL,
  drug_annotation = NULL
)
}
\arguments{
\item{df_metadata}{a data.table with metadata}

\item{cell_line_annotation}{optional data.table with cell line annotations;
if NULL (default), annotations are looked up from gDRinternal or gDRtestData}

\item{drug_annotation}{optional data.table with drug annotations;
if NULL (default), annotations are looked up from gDRinternal or gDRtestData}
}
\value{
a data.table with cleaned metadata
}
\description{
Cleanup a data.table with metadata
}
\details{
Adds annotations and check whether user provided correct input data.
}
\examples{

df <- data.table::data.table(
  clid = "CELL_LINE",
  Gnumber = "DRUG_1",
  Concentration = c(0, 1),
  Duration = 72
)
cleanup_df <- cleanup_metadata(df)

}
\keyword{utils}
