% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type.R
\name{collapse_drugs}
\alias{collapse_drugs}
\title{avoid \code{vehicle} in drug columns with low ordinality.
For example, if \code{drug = vehicle} and \code{drug_2 = drugA}, change to \code{drug = drugA} and \code{drug_2 = vehicle}.
propagate to the other columns accordingly (concentration, drug_moa, drug_name)}
\usage{
collapse_drugs(df)
}
\description{
avoid \code{vehicle} in drug columns with low ordinality.
For example, if \code{drug = vehicle} and \code{drug_2 = drugA}, change to \code{drug = drugA} and \code{drug_2 = vehicle}.
propagate to the other columns accordingly (concentration, drug_moa, drug_name)
}
