% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_se_to_raw_data.R
\name{convert_mae_to_raw_data}
\alias{convert_mae_to_raw_data}
\title{Transform mae into raw data}
\usage{
convert_mae_to_raw_data(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object with SummarizedExperiments containing "RawTreated" and "Controls" assays}
}
\value{
data.frame with raw data
}
\description{
Transform mae into raw data
}
\examples{
mae <- get_synthetic_data("finalMAE_small.RDS")
convert_mae_to_raw_data(mae)

}
