% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_se_to_raw_data.R
\name{convert_se_to_raw_data}
\alias{convert_se_to_raw_data}
\title{Transform se into raw_data}
\usage{
convert_se_to_raw_data(se)
}
\arguments{
\item{se}{SummarizedExperiment object with "RawTreated" and "Controls" assays}
}
\value{
data.frame with raw data
}
\description{
Transform se into raw_data
}
