% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_SE.R
\name{create_SE}
\alias{create_SE}
\title{Create a SummarizedExperiment object}
\usage{
create_SE(
  df_,
  readout = "ReadoutValue",
  control_mean_fxn = function(x) {     mean(x, trim = 0.25) },
  nested_keys = c("Barcode", gDRutils::get_identifier("masked_tag")),
  override_untrt_controls = NULL
)
}
\arguments{
\item{df_}{data.frame of raw drug response data containing both treated and untreated values.}

\item{readout}{string of the name containing the cell viability readout values.}

\item{control_mean_fxn}{function indicating how to average controls.
Defaults to \code{mean(x, trim = 0.25)}.}

\item{nested_keys}{character vector of column names to include in the data.frames in the assays of the resulting \code{SummarizedExperiment} object. 
Defaults to \code{c("Barcode", gDRutils::get_identifier("masked_tag"))}.}

\item{override_untrt_controls}{named list containing defining factors in the treatments.
Defaults to \code{NULL}.}
}
\value{
A \linkS4class{SummarizedExperiment} object containing two asssays.
Treated readouts will live in an assay called \code{"RawTreated"},
and reference readouts live in an assay called \code{"Controls"}.
\code{rownames} and \code{colnames} are made up of available metadata on treatments and conditions 
and is pasted together.
}
\description{
Create a SummarizedExperiment object from a data.frame, where the data.frame contains treatments on rows,
and conditions on columns.
}
\details{
This is most commonly used in preparation for downstream normalization.
}
\seealso{
Other runDrugResponseProcessingPipelineFxns: 
\code{\link{average_SE}()},
\code{\link{fit_SE}()},
\code{\link{normalize_SE}()},
\code{\link{runDrugResponseProcessingPipeline}()}
}
\concept{runDrugResponseProcessingPipelineFxns}
