% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_model.character}
\alias{data_model.character}
\title{Detect model of data from experiment name}
\usage{
\method{data_model}{character}(x)
}
\arguments{
\item{x}{character with experiment name}
}
\value{
string with the information of the raw data follows single-agent or combination data model
}
\description{
Detect model of data from experiment name
}
