% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_model.data.frame}
\alias{data_model.data.frame}
\title{Detect model of data in data.frame}
\usage{
\method{data_model}{data.frame}(x)
}
\arguments{
\item{x}{data.frame of raw drug response data containing both treated and untreated values.}
}
\value{
string with the information of the raw data follows single-agent or combination data model
}
\description{
Detect model of data in data.frame
}
