% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_model.data.table}
\alias{data_model.data.table}
\title{Detect model of data in data.table}
\usage{
\method{data_model}{data.table}(x)
}
\arguments{
\item{x}{data.table of raw drug response data
containing both treated and untreated values.}
}
\value{
string with the information of the raw data follows single-agent or
combination data model
}
\description{
Detect model of data in data.table
}
