% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_model}
\alias{data_model}
\title{Detect model of data}
\usage{
data_model(df_)
}
\arguments{
\item{df_}{data.frame of raw drug response data containing both treated and untreated values.}
}
\value{
string with the information of the raw data follows single-agent or combo data model
}
\description{
Detect model of data
}
