% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations-isobolograms.R
\name{define_matrix_grid_positions}
\alias{define_matrix_grid_positions}
\title{Define matrix grid positions}
\usage{
define_matrix_grid_positions(conc1, conc2)
}
\arguments{
\item{conc1}{drug_1 concentration}

\item{conc2}{drug_2 concentration}
}
\value{
list with axis grid positions
}
\description{
Define matrix grid positions
}
\details{
drug_1 is diluted along the rows as the y-axis and
drug_2 is diluted along the columns and will be the x-axis.
}
