% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_cores}
\alias{detect_cores}
\title{Function for detection of available cores used in parallel computing}
\usage{
detect_cores()
}
\value{
number of available cores
}
\description{
Function for detection of available cores used in parallel computing
}
