% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_treated_and_untreated_conditions.R
\name{.assign_treated_and_untreated_conditions}
\alias{.assign_treated_and_untreated_conditions}
\title{.assign_treated_and_untreated_conditions}
\usage{
.assign_treated_and_untreated_conditions(
  df_,
  drugname_col = gDRutils::get_env_identifiers("drugname")
)
}
\arguments{
\item{df_}{data.frame containing the drug name field.}

\item{drugname_col}{string of the column in \code{df_} corresponding to the drug name. 
Defaults to \code{gDRutils::get_env_identifiers("drugname")}..}
}
\value{
data.frame containing an additional factor column called \code{treated_untreated} specifying
whether the condition is a treated or untreated entry.
}
\description{
Assign the treated and untreated conditions to a data.frame
}
