% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{.create_combo_control}
\alias{.create_combo_control}
\title{Create combo control}
\usage{
.create_combo_control(nested_identifiers)
}
\arguments{
\item{nested_identifiers}{Character vector of the nested_identifiers for a given assay.}
}
\value{
\code{data.frame} to be nested in the assays.
}
\description{
Create combo control
}
\keyword{internal}
