% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df.R
\name{.map_references}
\alias{.map_references}
\title{Map references}
\usage{
.map_references(
  mat_elem,
  rowData_colnames = c(gDRutils::get_env_identifiers("duration"), paste0(c("drug",
    "drug_name", "drug_moa"), "3"))
)
}
\arguments{
\item{mat_elem}{data.table input containing experimental metadata and
row identifiers.}

\item{rowData_colnames}{character vector of variables (column names) used
to identify and map reference treatments.}
}
\value{
list
}
\description{
Map references
}
\keyword{internal}
\keyword{map_df}
