% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_curves.R
\name{.set_invalid_fit_params}
\alias{.set_invalid_fit_params}
\title{Set fit parameters for an invalid fit.}
\usage{
.set_invalid_fit_params(out, norm_values)
}
\arguments{
\item{out}{Named list of fit parameters.}

\item{norm_values}{Numeric vector used to estimate an \code{xc50} value.}
}
\value{
Modified named list of fit parameters.
}
\description{
Set fit parameters for an invalid fit.
}
\examples{
.set_invalid_fit_params(list(), norm_values = rep(0.3, 6))

}
\keyword{fit_curves}
