% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations-helpers.R
\name{.standardize_conc}
\alias{.standardize_conc}
\title{Standardize concentration values.}
\usage{
.standardize_conc(conc)
}
\arguments{
\item{conc}{numeric vector of the concentrations}
}
\value{
vector of standardized concentrations
}
\description{
Standardize concentration values.
}
\details{
If no \code{conc} are passed, \code{NULL} is returned.
}
\examples{
concs <- 10 ^ (seq(-1, 1, 0.9))
.standardize_conc(concs)

}
\keyword{utils}
