% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type.R
\name{duplicate_single_agent}
\alias{duplicate_single_agent}
\title{Duplicate single agent data to be visible also as Drug_2 in combo matrix}
\usage{
duplicate_single_agent(df_)
}
\arguments{
\item{df}{data.frame of raw drug response data containing both treated and untreated values}
}
\value{
data.frame of raw drug response data with additional duplicated single-agent data
}
\description{
Duplicate single agent data to be visible also as Drug_2 in combo matrix
}
\author{
Bartosz Czech <bartosz.czech@contractors.roche.com>
}
