% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_utils.R
\name{extrapolate_references}
\alias{extrapolate_references}
\title{Extrapolate references}
\usage{
extrapolate_references(ref_df, ref_conc)
}
\arguments{
\item{ref_df}{reference data data frame}

\item{ref_conc}{reference concentration values}
}
\value{
corrected prediction
}
\description{
Extrapolate references
}
