% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SE.combinations.R
\name{fit_SE.combinations}
\alias{fit_SE.combinations}
\title{fit_SE for combination screens}
\usage{
fit_SE.combinations(
  se,
  series_identifiers = NULL,
  normalization_types = c("RV", "GR"),
  averaged_assay = "Averaged"
)
}
\arguments{
\item{se}{\code{SummarizedExperiment} object with a BumpyMatrix assay containing averaged data.}

\item{series_identifiers}{character vector of the column names in the nested \code{DFrame}
corresponding to nested identifiers.
whose combination represents a unique series for which to fit curves.}

\item{conc_margin}{margin for calculation and plots as fold-change over highest test conc for calculation}

\item{log2_pos_offset}{max offset for conc}

\item{norm_types}{character vector of normalization types used for calculating combo matrix}
}
\value{
A code{SummarizedExperiment} object with an additional assay containing the combination metrics.
}
\description{
Perform fittings for combination screens.
}
\details{
This function assumes that the combination is set up with both concentrations nested in the assay.
}
