% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SE.R
\name{fit_SE}
\alias{fit_SE}
\title{fit_SE}
\usage{
fit_SE(avgSE, studyConcThresh = 4)
}
\arguments{
\item{avgSE}{a SummarizedExperiment with averaged and normalized assays}

\item{studyConcThresh}{a numeric with study concentration threshold (4 by default)}
}
\value{
a SummarizedExperiment with additional assay with metrics
}
\description{
Calculate metrics for DR data
}
