% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\docType{package}
\name{gDRcore}
\alias{gDRcore}
\title{'gDRcore - a package with core functionality'}
\description{
gDRcore uses `BiocParallel` package to make computation parallel using all available cores.
It is possible to switch into sequential mode by modifying gDRcore settings.
If you want to use a sequential approach you can easily run `options(MulticoreParam = SerialParam())`
to turn off parallel computing mode. For other available settings please visit the documentation
of `BiocParallel` package:
https://bioconductor.org/packages/release/bioc/html/BiocParallel.html
}
