% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateCodilution}
\alias{generateCodilution}
\title{generateCodilution}
\usage{
generateCodilution(cell_lines, drugs, save = TRUE)
}
\value{
data.table with raw input data or MAE with processed data
}
\description{
generateCodilution
}
\keyword{internal}
