% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateComboManyDrugs}
\alias{generateComboManyDrugs}
\title{generateComboManyDrugs}
\usage{
generateComboManyDrugs(cell_lines, drugs, save = TRUE)
}
\value{
data.frame with raw input data or MAE with processed data
}
\description{
generateComboManyDrugs
}
\examples{

cell_lines <- gDRtestData::create_synthetic_cell_lines()
drugs <- gDRtestData::create_synthetic_drugs()
generateComboManyDrugs(cell_lines, drugs)

}
\keyword{internal}
