% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateComboNoNoiseData2}
\alias{generateComboNoNoiseData2}
\title{generateComboNoNoiseData2}
\usage{
generateComboNoNoiseData2(cell_lines, drugs, save = TRUE)
}
\value{
data.frame with raw input data or MAE with processed data
}
\description{
generateComboNoNoiseData2
}
\examples{

cell_lines <- gDRtestData::create_synthetic_cell_lines()
drugs <- gDRtestData::create_synthetic_drugs()
generateComboNoNoiseData2(cell_lines, drugs)

}
\keyword{internal}
