% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateComboNoNoiseData3}
\alias{generateComboNoNoiseData3}
\title{generateComboNoNoiseData3}
\usage{
generateComboNoNoiseData3(cell_lines, drugs, save = TRUE)
}
\value{
data.frame with raw input data or MAE with processed data
}
\description{
generateComboNoNoiseData3
}
\keyword{internal}
