% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateManyLinesData}
\alias{generateManyLinesData}
\title{generateManyLinesData}
\usage{
generateManyLinesData(cell_lines, drugs, save = TRUE)
}
\value{
data.frame with raw input data or MAE with processed data
}
\description{
generateManyLinesData
}
\examples{

cell_lines <- create_synthetic_cell_lines()
drugs <- create_synthetic_drugs()
generateManyLinesData(cell_lines, drugs)

}
\keyword{internal}
