% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateMediumData}
\alias{generateMediumData}
\title{generateMediumData}
\usage{
generateMediumData(cell_lines, drugs, save = TRUE)
}
\value{
data.frame with raw input data or MAE with processed data
}
\description{
generateMediumData
}
\keyword{internal}
