% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateTripleComboMatrix}
\alias{generateTripleComboMatrix}
\title{generateTripleComboMatrix}
\usage{
generateTripleComboMatrix(cell_lines, drugs, save = TRUE)
}
\value{
data.frame with raw input data or MAE with processed data
}
\description{
generateTripleComboMatrix
}
\examples{

cell_lines <- create_synthetic_cell_lines()
drugs <- create_synthetic_drugs()
generateTripleComboMatrix(cell_lines, drugs)

}
\keyword{internal}
