% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_assays_per_pipeline_step}
\alias{get_assays_per_pipeline_step}
\title{get info about created/present assays in SE at the given pipeline step}
\usage{
get_assays_per_pipeline_step(
  step,
  data_model,
  status = c("created", "present")
)
}
\arguments{
\item{step}{string with pipeline step}

\item{data_model}{single-agent vs combination}

\item{status}{string return vector of assays created or present at the given step?}
}
\description{
get info about created/present assays in SE at the given pipeline step
}
