% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{get_cellline_annotation_from_dt}
\alias{get_cellline_annotation_from_dt}
\title{Retrieve the cell line annotation from the annotated dt input}
\usage{
get_cellline_annotation_from_dt(dt)
}
\arguments{
\item{dt}{annotated data.table}
}
\value{
data.table with cell line annotation
}
\description{
Retrieve the cell line annotation from the annotated dt input
}
\examples{
dt <- data.table::data.table(Gnumber = "A",
clid = "CL123",
CellLineName = "cl name",
Tissue = "Bone",
parental_identifier = "some cl",
subtype = "cortical",
ReferenceDivisionTime = 5)
get_cellline_annotation_from_dt(dt)
}
\keyword{annotation}
