% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_default_nested_identifiers}
\alias{get_default_nested_identifiers}
\alias{get_default_nested_identifiers.data.frame}
\alias{get_default_nested_identifiers.SummarizedExperiment}
\title{Get default nested identifiers}
\usage{
get_default_nested_identifiers(x, data_model = NULL)

\method{get_default_nested_identifiers}{data.frame}(x, data_model = NULL)

\method{get_default_nested_identifiers}{SummarizedExperiment}(x, data_model = NULL)
}
\arguments{
\item{x}{data.frame with raw data or SummarizedExperiment object with gDR assays}

\item{data_model}{single-agent vs combination}
}
\value{
vector of nested identifiers
}
\description{
Get default nested identifiers
}
