% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{get_drug_annotation_from_dt}
\alias{get_drug_annotation_from_dt}
\title{Retrieve the drug annotation from the annotated dt input}
\usage{
get_drug_annotation_from_dt(dt)
}
\arguments{
\item{dt}{annotated data.table}
}
\value{
data.table with drug annotation
}
\description{
Retrieve the drug annotation from the annotated dt input
}
\examples{
dt <- data.table::data.table(Gnumber = "A",
DrugName = "drugA",
drug_moa = "drug_moa_A")
get_drug_annotation_from_dt(dt)
}
\keyword{annotation}
