% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mae_from_intermediate_data}
\alias{get_mae_from_intermediate_data}
\title{get mae dataset from intermediate data}
\usage{
get_mae_from_intermediate_data(data_dir)
}
\arguments{
\item{data_dir}{directory with intermediate data}
}
\description{
get mae dataset from intermediate data
}
