% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nested_identifiers}
\alias{get_nested_identifiers}
\title{Get default nested identifiers}
\usage{
get_nested_identifiers(df_)
}
\arguments{
\item{df_}{data.frame of raw drug response data containing both treated and untreated values.}
}
\value{
vector of nested identifiers
}
\description{
Get default nested identifiers
}
