% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_SE_nested_identifiers}
\alias{get_SE_nested_identifiers}
\title{Get se nested identifiers}
\usage{
get_SE_nested_identifiers(
  se,
  assayName = tail(SummarizedExperiment::assayNames(se), 1)
)
}
\arguments{
\item{se}{SummarizedExperiment}

\item{assayName}{assay name used for checking nested identifiers}
}
\value{
vector of nested identifiers
}
\description{
Get se nested identifiers
}
