% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR_curve_fit.R
\name{GRlogisticFit}
\alias{GRlogisticFit}
\title{Actual fitting function}
\usage{
GRlogisticFit(log10concs, GRvalues, upper_GR = 1, force = FALSE,
  cap = FALSE)
}
\arguments{
\item{log10concs}{concentrations}

\item{GRvalues}{values}

\item{upper_GR}{=1 by default}

\item{force}{use signifcance or not}

\item{cap}{enforce upper_GR}
}
\value{
vector of values
}
\description{
\code{GRlogisticFit} returns fit parameters
}
\details{
returns fit parameters
}
\examples{
sum(1:10)
}
