% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_utils.R
\name{infer_control_df}
\alias{infer_control_df}
\title{Create a control dataframe for a treatment-cell line combination with inferred data.}
\usage{
infer_control_df(
  df_,
  infer_concentration,
  control_cols,
  control_mean_fxn = function(x) mean(x, trim = 0.25),
  out_col_name
)
}
\arguments{
\item{df_}{data.frame}

\item{infer_concentration}{concentration at which the \code{out_col_name} should be evaluated.}

\item{control_cols}{character vector of columns to include in the resulting control data.frame.}

\item{control_mean_fxn}{function indicating how to average controls.
Defaults to \code{mean(x, trim = 0.25)}.}

\item{out_col_name}{string of the output readout that will replace \code{CorrectedReadout}.}
}
\value{
data.frame of values aggregated by
}
\description{
Create an aggregated control data.frame based on inferring a given concentration usign a sigmoidal fit.
}
\examples{

}
