% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{is_equal_concentration}
\alias{is_equal_concentration}
\title{Equal concentrations to ndigit significant digits in log10 domain}
\usage{
is_equal_concentration(x, y, ndigit = 3)
}
\arguments{
\item{x}{value to be tested.}

\item{y}{value to be tested.}

\item{ndigit}{number of significant digits (default = 3).}
}
\value{
boolean
}
\description{
Equal concentrations to ndigit significant digits in log10 domain
}
