% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDrugResponseProcessingPipeline.R
\name{is_preceding_step}
\alias{is_preceding_step}
\title{check if the given step is preceding the step chosen in the partial run}
\usage{
is_preceding_step(current_step, start_from, steps = get_pipeline_steps())
}
\arguments{
\item{current_step, }{string with the step to be evaluated}

\item{start_from}{string indicating the pipeline step from which partial run should be launched}

\item{steps}{charvect with all available steps}
}
\value{
logical
}
\description{
check if the given step is preceding the step chosen in the partial run
}
\keyword{internal}
