% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations-helpers.R
\name{map_ids_to_fits}
\alias{map_ids_to_fits}
\title{Get predicted values for a given fit and input.}
\usage{
map_ids_to_fits(pred, match_col, fittings, fitting_id_col)
}
\arguments{
\item{pred}{numeric vector for which you want predictions.}

\item{match_col}{vector to match on \code{fittings} to get the correct fit.}

\item{fittings}{data.frame of fit metrics.}

\item{fitting_id_col}{string of the column name in \code{fittings} that 
should be used to match with \code{match_col} .}
}
\value{
numeric vector of predicted values given \code{pred} inputs and 
\code{fittings} values.
}
\description{
Map fittings to identifiers and compute the predicted values for 
corresponding fits.
}
