% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_SE.R
\name{map_SE}
\alias{map_SE}
\title{map_SE}
\usage{
map_SE(normSE, ctrlSE, row_endpoint_value_filter, Keys, T0 = FALSE)
}
\arguments{
\item{normSE}{a SummarizedExperiment with normalization assaay}

\item{ctrlSE}{a SummarizedExperiment object with information for controls}

\item{row_endpoint_value_filter}{an array with key values for end points}

\item{Keys}{a list of all identified keys}

\item{T0}{a logical indicating if the mapping should be performer for Time=0 (FALSE by default)}
}
\value{
a list of mapping
}
\description{
Perfmorm mapping for normalization
}
