% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df.R
\name{map_untreated}
\alias{map_untreated}
\title{Identify untreated rows based on Drug treatment alone}
\usage{
map_untreated(mat_elem)
}
\arguments{
\item{mat_elem}{data.table input containing drug name or drug identifier
columns.}
}
\value{
logical vector
}
\description{
Identify untreated rows based on Drug treatment alone
}
\keyword{map_df}
