% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{merge_data}
\usage{
merge_data(
  manifest,
  treatments,
  data,
  cell_line_annotation = NULL,
  drug_annotation = NULL
)
}
\arguments{
\item{manifest}{a data.table with a manifest info}

\item{treatments}{a data.table with a treaatments info}

\item{data}{a data.table with a raw data info}

\item{cell_line_annotation}{optional data.table with cell line annotations;
if NULL (default), annotations are looked up from gDRinternal or gDRtestData}

\item{drug_annotation}{optional data.table with drug annotations;
if NULL (default), annotations are looked up from gDRinternal or gDRtestData}
}
\value{
a data.table with merged data and metadata.
}
\description{
Merge all the input data into a single data.table
}
\examples{
td <- gDRimport::get_test_data()
l_tbl <- gDRimport::load_data(
  manifest_file = gDRimport::manifest_path(td),
  df_template_files = gDRimport::template_path(td),
  results_file = gDRimport::result_path(td)
)
merge_data(
  l_tbl$manifest,
  l_tbl$treatments,
  l_tbl$data
)

}
\keyword{merge_data}
