% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_SE.R
\name{normalize_SE}
\alias{normalize_SE}
\title{normalize_SE}
\usage{
normalize_SE(
  se,
  control_assay = "Controls",
  raw_treated_assay = "RawTreated",
  normalized_assay = "Normalized",
  ref_GR_assay = "RefGRvalue",
  ref_RV_assay = "RefRelativeViability",
  ndigit_rounding = 4
)
}
\arguments{
\item{se}{\code{BumpyMatrix} object with assays \code{"RawTreated"} and \code{"Controls"}.}

\item{control_assay}{string containing the name of the assay representing the controls in the \code{se}.
Defaults to \code{"Controls"}.}

\item{raw_treated_assay}{string containing the name of the assay representing the raw treated data in the \code{se}.
Defaults to \code{"RawTreated"}.}

\item{normalized_assay}{string of the assay name containing the normalized data.
Defaults to \code{"Normalized"}.}

\item{ref_GR_assay}{string of the name of the reference GR assay in the \linkS4class{SummarizedExperiment}.}

\item{ref_RV_assay}{string of the name of the reference
Relative Viability assay in the \linkS4class{SummarizedExperiment}.}

\item{ndigit_rounding}{integer specifying number of digits of rounding during calculations.
Defaults to \code{4}.}
}
\value{
\code{BumpyMatrix} object with a new assay named \code{"Normalized"}containing \code{DataFrame}s 
holding \code{RelativeViability} and \code{GRvalue}, as well as new assays named \code{RefRelativeViability},
\code{RefGRvalue}, and \code{DivisionTime} values.
}
\description{
Normalize drug response data from treated and untreated pairings.
}
\seealso{
Other runDrugResponseProcessingPipelineFxns: 
\code{\link{average_SE}()},
\code{\link{create_SE}()},
\code{\link{fit_SE}()},
\code{\link{runDrugResponseProcessingPipeline}()}
}
\concept{runDrugResponseProcessingPipelineFxns}
