% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parallel}
\alias{parallel}
\title{Function for detection of available cores used in parallel computing}
\usage{
parallel(x, FUN, ...)
}
\value{
number of available cores
}
\description{
Function for detection of available cores used in parallel computing
}
