% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parallelize}
\alias{parallelize}
\title{Function for detection of available cores used in parallelize computing}
\usage{
parallelize(x, FUN, ...)
}
\value{
number of available cores
}
\description{
Function for detection of available cores used in parallelize computing
}
