% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDrugResponseProcessingPipeline.R
\name{prepare_input.data.frame}
\alias{prepare_input.data.frame}
\title{Prepare input data common for all experiments}
\usage{
\method{prepare_input}{data.frame}(
  x,
  nested_confounders = gDRutils::get_env_identifiers("barcode"),
  nested_identifiers_l = .get_default_nested_identifiers()
)
}
\arguments{
\item{x}{data.frame with raw data}
}
\description{
Current steps
- refining nested confounders
- refining nested identifiers
- splitting df_ into (per experiment) df_list
}
