% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input.R
\name{prepare_input}
\alias{prepare_input}
\title{Prepare input data common for all experiments}
\usage{
prepare_input(x, ...)
}
\arguments{
\item{x}{data.frame with raw data or MAE object with dose-reponse data}

\item{...}{additional parameters}
}
\value{
list of input data
}
\description{
Current steps
- refining nested confounders
- refining nested identifiers
- splitting df_ into (per experiment) df_list
}
\examples{
dataDir <- system.file("extdata", "data1", package = "gDRimport")
manifest <- list.files(dataDir, pattern = "manifest", full.names = TRUE)
template <- list.files(dataDir, pattern = "Template", full.names = TRUE)
raw_data <- list.files(dataDir, pattern = "^RawData", full.names = TRUE)
l_data <- gDRimport::load_data(manifest, template, raw_data)
df_ <- gDRcore::merge_data(
  l_data$manifest, 
  l_data$treatments, 
  l_data$data
)
nested_confounders = intersect(
  names(df_), 
  gDRutils::get_env_identifiers("barcode")
)
prepare_input(df_, nested_confounders, NULL)

}
