% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDrugResponseProcessingPipeline.R
\name{read_intermediate_data}
\alias{read_intermediate_data}
\title{read intermediate data for the given experiment and step to qs file}
\usage{
read_intermediate_data(path, step, experiment)
}
\arguments{
\item{path}{string with the input directory of the qs file}

\item{step, }{string with the step name}

\item{experiment}{string with the experiment name}
}
\value{
se
}
\description{
read intermediate data for the given experiment and step to qs file
}
\keyword{internal}
