% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{remove_drug_batch}
\alias{remove_drug_batch}
\title{Remove batch from Gnumber}
\usage{
remove_drug_batch(drug)
}
\arguments{
\item{drug}{drug name}
}
\value{
Gnumber without a batch
}
\description{
Remove batch from Gnumber
}
