% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_concentration}
\alias{replace_concentration}
\title{Replace concentrations in x by values in y if matching to the ndigit significant digits in log10 domain}
\usage{
replace_concentration(x, y, ndigit = 2)
}
\arguments{
\item{x}{value to change if equal to values in y.}

\item{y}{reference values.}

\item{ndigit}{number of significant digits (default = 2).}
}
\value{
adjusted x values
}
\description{
Replace concentrations in x by values in y if matching to the ndigit significant digits in log10 domain
}
