% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_concentration}
\alias{round_concentration}
\title{Round concentration to ndigit significant digits}
\usage{
round_concentration(x, ndigit = 3)
}
\arguments{
\item{x}{value to be rounded.}

\item{ndigit}{number of significant digits (default = 4).}
}
\value{
rounded x
}
\description{
Round concentration to ndigit significant digits
}
\examples{
round_concentration(x = c(0.00175,0.00324,0.0091), ndigit = 1)

}
