% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDrugResponseProcessingPipeline.R
\name{save_intermediate_data}
\alias{save_intermediate_data}
\title{save intermediate data for the given experiment and step to qs file}
\usage{
save_intermediate_data(path, step, experiment, se)
}
\arguments{
\item{path}{string with the save directory for the qs file}

\item{step, }{string with the step name}

\item{experiment}{string with the experiment name}

\item{se}{output se}
}
\description{
save intermediate data for the given experiment and step to qs file
}
\keyword{internal}
