% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type.R
\name{split_raw_data}
\alias{split_raw_data}
\title{Split raw data into list based on the data types}
\usage{
split_raw_data(df, type_col = "type")
}
\arguments{
\item{df}{data.frame of raw drug response data containing both treated and 
untreated values with column specified in `type_col` argument.}

\item{type_col}{string with column names in `df` with info about data type.
Defaults to \code{"type"}.}
}
\value{
list with split data based on its data type
}
\description{
Split raw data into list based on the data types
}
\examples{
df_layout <- merge(cell_lines[7:8, ], drugs[c(4:6), ], by = NULL)
df_layout <- gDRtestData::add_data_replicates(df_layout)
df_layout <- gDRtestData::add_concentration(
  df_layout, 
  concentrations = 10 ^ (seq(-3, .5, .5))
)
df_2 <- merge(
  cell_lines[cell_lines$clid \%in\% df_layout$clid, ], 
  drugs[c(21, 26), ], 
  by = NULL
)
df_2 <- gDRtestData::add_data_replicates(df_2)
df_2 <- gDRtestData::add_concentration(
  df_2, 
  concentrations = 10 ^ (seq(-3, .5, .5))
)
colnames(df_2)[colnames(df_2) \%in\% c(colnames(drugs), "Concentration")] <- 
  paste0(
    colnames(df_2)[colnames(df_2) \%in\% c(colnames(drugs), "Concentration")], 
    "_2"
  )
df_layout_2 <- merge(df_layout, df_2)
df_merged_data <- gDRtestData::generate_response_data(df_layout_2, 0)
df <- identify_data_type(df_merged_data)
split_raw_data(df)

}
\author{
Bartosz Czech <bartosz.czech@contractors.roche.com>
}
