% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.utils.R
\name{test_se}
\alias{test_se}
\title{test_se}
\usage{
test_se(se, lRef)
}
\arguments{
\item{se}{a SummarizedExperiment object with the calculated assay}

\item{lRef}{a list of reference datasets}
}
\value{

}
\description{
Compare all the assays in the SummarizedExperiment object with the reference
}
