% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.utils.R
\name{test_synthetic_data2}
\alias{test_synthetic_data2}
\title{Test that the data is consistent after moving the RefReadout out of the create_and_normalize_SE logic.}
\usage{
test_synthetic_data2(
  original,
  long_df,
  dataName,
  asys = c("Normalized", "Averaged")
)
}
\arguments{
\item{original}{original MAE assay}

\item{long_df}{dataset in data frame format}

\item{dataName}{dataset name}

\item{asys}{names of assays}
}
\description{
Test that the data is consistent after moving the RefReadout out of the create_and_normalize_SE logic.
}
