% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_data_models_availability}
\alias{validate_data_models_availability}
\title{Validate availability of data models}
\usage{
validate_data_models_availability(d_types, s_d_models)
}
\arguments{
\item{d_types}{character vector with experiment names in \code{MultiAssayExperiment} object}

\item{s_d_models}{character vector with names of supported experiment}
}
\description{
Validate availability of data models
}
\keyword{internal}
