% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{add_CellLine_annotation}
\alias{add_CellLine_annotation}
\title{add_CellLine_annotation}
\usage{
add_CellLine_annotation(
  dt_metadata,
  DB_cellid_header = "cell_line_identifier",
  DB_cell_annotate = c("cell_line_name", "primary_tissue", "doubling_time",
    "parental_identifier", "subtype"),
  fname = "cell_lines.csv",
  fill = "unknown",
  annotationPackage = if ("gDRinternal" \%in\% .packages(all.available = TRUE)) {
    
    "gDRinternal"
 } else {
     "gDRtestData"
 }
)
}
\arguments{
\item{dt_metadata}{data.table with metadata}

\item{DB_cellid_header}{string with colnames with cell line identifier
in the annotation file}

\item{DB_cell_annotate}{character vector with mandatory colnames
used in the annotation file}

\item{fname}{string with file name with annotation}

\item{fill}{string indicating how unknown cell lines should be filled in the DB}

\item{annotationPackage}{string indication name of the package containing cellline annotation}
}
\value{
data.table with metadata with annotated cell lines
}
\description{
add cellline annotation to a data.table with metadata
}
\details{
The logic of adding celline annotation for dt_metadata based on
the annotation file stored in gDRtestData. Other fields are set as "unknown".
This approach will be corrected once we will implement final solution
for adding cell lines.
}
\examples{

add_CellLine_annotation(
  data.table::data.table(
    clid = "123", 
    CellLineName = "name of the cell line")
)

}
\keyword{annotation}
