% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{add_Drug_annotation}
\alias{add_Drug_annotation}
\title{add_Drug_annotation}
\usage{
add_Drug_annotation(
  df_metadata,
  fname = "drugs.csv",
  fill = "unknown",
  annotationPackage = if (requireNamespace("gDRinternalData", quietly = TRUE)) {
    
    "gDRinternalData"
 } else {
     "gDRtestData"
 }
)
}
\arguments{
\item{df_metadata}{data.frame with metadata}

\item{fname}{string with file name with annotation}

\item{fill}{string indicating how unknown cell lines should be filled in the DB}

\item{annotationPackage}{string indication name of the package containing drug annotation}
}
\value{
data.frame with metadata with annotated drugs
}
\description{
add drug annotation to a data.frame with metadata
}
\details{
The logic of adding drug annotation for df_metadata
based on the annotation file stored in gDRtestData.
}
\examples{
add_Drug_annotation(
  data.frame(
    Gnumber = "drug_id", 
    DrugName = "name of the drug")
)
}
