% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{add_Drug_annotation}
\alias{add_Drug_annotation}
\title{add_Drug_annotation}
\usage{
add_Drug_annotation(
  dt_metadata,
  fname = "drugs.csv",
  fill = "unknown",
  annotationPackage = if ("gDRinternal" \%in\% .packages(all.available = TRUE)) {
    
    "gDRinternal"
 } else {
     "gDRtestData"
 },
  externalSource = Sys.getenv("GDR_DRUG_ANNOTATION")
)
}
\arguments{
\item{dt_metadata}{data.table with metadata}

\item{fname}{string with file name with annotation}

\item{fill}{string indicating how unknown cell lines should be filled in the DB}

\item{annotationPackage}{string indication name of the package containing drug annotation}

\item{externalSource}{string with path to external file with annotation data; by default it checks
'GDR_DRUG_ANNOTATION' env var. This file should contain columns such as cell_line_identifier,
cell_line_name, primary_tissue, doubling_time, parental_identifier, and subtype}
}
\value{
data.table with metadata with annotated drugs
}
\description{
add drug annotation to a data.table with metadata
}
\details{
The logic of adding drug annotation for dt_metadata
based on the annotation file stored in gDRtestData.
}
\examples{
add_Drug_annotation(
  data.table::data.table(
    Gnumber = "drug_id", 
    DrugName = "name of the drug")
)
}
\keyword{annotation}
