% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_intermediate_data}
\alias{add_intermediate_data}
\title{add intermediate data (qs files) for given ma}
\usage{
add_intermediate_data(mae, data_dir, steps = get_pipeline_steps())
}
\arguments{
\item{mae}{mae with dose-response data}

\item{data_dir}{output directory}

\item{steps}{charvec with pipeline steps for which intermediate data should be saved}
}
\description{
add intermediate data (qs files) for given ma
}
