% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{calculate_combo_cotrt}
\alias{calculate_combo_cotrt}
\title{Calculate combo co-treatment}
\usage{
calculate_combo_cotrt(se)
}
\arguments{
\item{se}{a BumpyMatrix SE with drug response data}
}
\description{
Calculate combo co-treatment
}
