% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_combo_matrix.R
\name{calculate_combo_matrix}
\alias{calculate_combo_matrix}
\title{calculate_combo_matrix}
\usage{
calculate_combo_matrix(
  SE,
  conc_margin = 10^0.5,
  log2_pos_offset = log10(3)/2,
  norm_types = c("RelativeViability", "GRvalue")
)
}
\arguments{
\item{SE}{a BumpyMatrix SE with drug response data}

\item{conc_margin}{margin for calculation and plots as fold-change over highest test conc for calculation}

\item{log2_pos_offset}{max offset for conc}
}
\value{
data
}
\description{
calculate_combo_matrix
}
