% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_combo_matrix.R
\name{calculate_combo_matrix}
\alias{calculate_combo_matrix}
\title{calculate_combo_matrix}
\usage{
calculate_combo_matrix(
  se,
  series_identifiers,
  conc_margin = 10^0.5,
  log2_pos_offset = log10(3)/2,
  norm_types = c("RelativeViability", "GRvalue")
)
}
\arguments{
\item{se}{a BumpyMatrix SE with drug response data}

\item{series_identifiers}{character vector of the column names in \code{data.frame}
whose combination represents a unique series for which to fit curves.}

\item{conc_margin}{margin for calculation and plots as fold-change over highest test conc for calculation}

\item{log2_pos_offset}{max offset for conc}

\item{norm_types}{character vector of normalization types used for calculating combo matrix}
}
\value{
data
}
\description{
calculate_combo_matrix
}
