% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations-calculate_matrix_metric.R
\name{calculate_matrix_metric}
\alias{calculate_matrix_metric}
\alias{calculate_HSA}
\alias{calculate_Bliss}
\title{Calculate a metric for combination data.}
\usage{
calculate_HSA(sa1, series_id1, sa2, series_id2, metric)

calculate_Bliss(sa1, series_id1, sa2, series_id2, metric)
}
\arguments{
\item{sa1}{data.frame containing single agent data where id1 is \code{0}.
Columns of the data.frame include identifiers and the \code{metric} of interest.}

\item{series_id1}{String representing the column within \code{sa1} that represents id1.}

\item{sa2}{data.frame containing single agent data where id2 is \code{0}.
Columns of the data.frame include identifiers and the \code{metric} of interest.}

\item{series_id2}{String representing the column within \code{sa1} that represents id2.}

\item{metric}{String of the column specifying the metric of interest.}

\item{FXN}{Function to apply to the single-agent fits to calculate a metric.}
}
\value{
DataFrame containing a single row for every unique combination of the two series
identifiers and the corresponding calculated metric for each row.
}
\description{
Calculate a metric based off of single-agent values in combination screens.
}
\details{
\code{calculate_HSA} takes the minimum of the two single agents readouts.
\code{calculate_Bliss} performs Bliss additivity calculation based on the single agent effects,
defined as \code{1-x} for the corresponding normalization.
See https://www.sciencedirect.com/science/article/pii/S1359644619303460?via%3Dihub#tb0005
for more details.
}
