% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_model}
\alias{data_model}
\title{Detect model of data}
\usage{
data_model(x)
}
\arguments{
\item{x}{data.table with raw data or SummarizedExperiment object 
with gDR assays}
}
\value{
string with the information of the raw data follows single-agent or 
combination data model
}
\description{
Detect model of data
}
\examples{
data_model("single-agent")

}
