% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_conditions.R
\name{.assign_conditions}
\alias{.assign_conditions}
\title{.assign_conditions}
\usage{
.assign_conditions(df_, nested_identifiers)
}
\arguments{
\item{df_}{data.frame containing the \code{nested_identifiers}.}

\item{nested_identifiers}{Character vector of column names in \code{df_}
that will determine a condition type.}
}
\value{
data.frame containing an additional factor column called \code{"conditions"} specifying
whether the condition is a \code{"treated"}, \code{"untreated"}, or \code{"reference"} entry.
}
\description{
Assign the treated, untreated, and reference conditions to a data.frame
}
\details{
Conditions are considered \code{"untreated"} when both \code{nested_identifiers}
have a value of 0. When nested_identifiers are drugs, this represent entries where all drugs 
are at 0 concentration. Conditions are considered \code{"reference"} when only one
of the \code{nested_identifiers} is 0. In the context of drugs, this represents a single-agent 
response.
}
