% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df.R
\name{.get_untreated_tag_count}
\alias{.get_untreated_tag_count}
\title{Get the count of untreated tags per row}
\usage{
.get_untreated_tag_count(
  mat_elem,
  drug_identifier_keys = c("drug_name", "drug_name2", "drug_name3")
)
}
\arguments{
\item{mat_elem}{data.table input data frame}

\item{drug_identifier_keys}{character vector of keys to look up identifiers}
}
\value{
list containing ntag, num_cols, and valid_cols
}
\description{
Get the count of untreated tags per row
}
\keyword{internal}
