% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df.R
\name{.map_references}
\alias{.map_references}
\title{Map references}
\usage{
.map_references(mat_elem)
}
\description{
Map references
}
\details{
Using the given rownames, map the treated and reference conditions.
}
