% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations-helpers.R
\name{.standardize_conc}
\alias{.standardize_conc}
\title{Standardize concentration values.}
\usage{
.standardize_conc(conc)
}
\arguments{
\item{conc}{numeric vector of the concentrations}
}
\value{
vector of standardized concentrations
}
\description{
Standardize concentration values.
}
