% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SE.R
\name{fit_SE}
\alias{fit_SE}
\title{fit_SE}
\usage{
fit_SE(
  se,
  averaged_assay = "Averaged",
  ref_GR_assay = "RefGRvalue",
  ref_RV_assay = "RefRelativeViability",
  metrics_assay = "Metrics",
  n_point_cutoff = 4,
  range_conc = c(0.005, 5),
  force_fit = FALSE,
  pcutoff = 0.05,
  cap = 0.1,
  curve_type = c("GR", "RV")
)
}
\arguments{
\item{se}{a \linkS4class{SummarizedExperiment} with normalized, averaged assays.
Corresponding assay names are specified by \code{averaged_assay}, \code{ref_GR_assay}, and \code{ref_RV_assy}.}

\item{averaged_assay}{string of the name of the averaged assay in the \linkS4class{SummarizedExperiment}.
Defaults to \code{"Averaged"}.}

\item{ref_GR_assay}{string of the name of the reference GR assay in the \linkS4class{SummarizedExperiment}.
Defaults to \code{"RefGRvalue"}.}

\item{ref_RV_assay}{string of the name of the reference Relative Viability assay in the \linkS4class{SummarizedExperiment}.
Defaults to \code{"RefRelativeViability"}.}

\item{metrics_assay}{string of the name of the metrics assay to output in the returned \linkS4class{SummarizedExperiment}.
Defaults to \code{"Metrics"}.}

\item{n_point_cutoff}{integer of how many points should be considered the minimum required to try to fit a curve.
Defaults to \code{4}.}

\item{range_conc}{numeric vector of length 2 indicating the lower and upper concentration ranges.
Defaults to \code{c(5e-3, 5)}. See details.}

\item{force_fit}{boolean indicating whether or not to force a constant fit.
Defaults to \code{FALSE}.}

\item{pcutoff}{numeric of pvalue significance threshold above or equal to which to use a constant fit.
Defaults to \code{0.05}.}

\item{cap}{numeric value capping \code{norm_values} to stay below (\code{x_0} + cap).
Defaults to \code{0.1}.}

\item{curve_type}{character vector of types of curves to fit.
Defaults to \code{c("GR", "RV")}.}
}
\value{
the original \linkS4class{SummarizedExperiment} with an additional assay 
containing fitting metrics named \code{metrics_assay}.
}
\description{
Fit curves and obtain fit metrics from normalized, averaged drug response data.
}
\seealso{
Other runDrugResponseProcessingPipelineFxns: 
\code{\link{average_SE}()},
\code{\link{create_SE}()},
\code{\link{normalize_SE}()},
\code{\link{runDrugResponseProcessingPipeline}()}
}
\concept{runDrugResponseProcessingPipelineFxns}
