% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDRcore-package.R
\docType{package}
\name{gDRcore-package}
\alias{gDRcore}
\alias{gDRcore-package}
\title{gDRcore: Processing functions and interface to process and analyze drug dose-response data}
\value{
package help page
}
\description{
This package contains core functions to process and analyze drug response data. The package provides tools for normalizing, averaging, and calculation of gDR metrics data. All core functions are wrapped into the pipeline function allowing analyzing the data in a straightforward way.
}
\note{
To learn more about functions start with \code{help(package = "gDRcore")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gdrplatform/gDRcore}
  \item \url{https://gdrplatform.github.io/gDRcore/}
  \item Report bugs at \url{https://github.com/gdrplatform/gDRcore/issues}
}

}
\author{
\strong{Maintainer}: Arkadiusz Gladki \email{gladki.arkadiusz@gmail.com} (\href{https://orcid.org/0000-0002-7059-6378}{ORCID})

Authors:
\itemize{
  \item Bartosz Czech \email{bartosz.czech@contractors.roche.com} (\href{https://orcid.org/0000-0002-9908-3007}{ORCID})
  \item Marc Hafner (\href{https://orcid.org/0000-0003-1337-7598}{ORCID})
  \item Pawel Piatkowski
  \item Natalia Potocka
  \item Dariusz Scigocki
  \item Janina Smola
  \item Sergiu Mocanu
  \item Marcin Kamianowski
  \item Allison Vuong
}

}
\keyword{internal}
