% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateCodilutionSmall}
\alias{generateCodilutionSmall}
\title{generateCodilutionSmall}
\usage{
generateCodilutionSmall(cell_lines, drugs, save = TRUE)
}
\value{
data.table with raw input data or MAE with processed data
}
\description{
generateCodilutionSmall
}
\keyword{internal}
