% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateComboMatrix}
\alias{generateComboMatrix}
\title{generateComboMatrix}
\usage{
generateComboMatrix(cell_lines, drugs, save = TRUE)
}
\value{
data.frame with raw input data or MAE with processed data
}
\description{
generateComboMatrix
}
\keyword{internal}
