% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateComboMatrixSmall}
\alias{generateComboMatrixSmall}
\title{generateComboMatrixSmall}
\usage{
generateComboMatrixSmall(cell_lines, drugs, save = TRUE)
}
\value{
data.frame with raw input data or MAE with processed data
}
\description{
generateComboMatrixSmall
}
\examples{

cell_lines <- create_synthetic_cell_lines()
drugs <- create_synthetic_drugs()
generateComboMatrixSmall(cell_lines, drugs)

}
\keyword{internal}
