% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateComboNoNoiseData}
\alias{generateComboNoNoiseData}
\title{generateComboNoNoiseData}
\usage{
generateComboNoNoiseData(cell_lines, drugs, save = TRUE)
}
\value{
data.frame with raw input data or MAE with processed data
}
\description{
generateComboNoNoiseData
}
\keyword{internal}
