% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateNoiseRawData}
\alias{generateNoiseRawData}
\title{generateNoiseRawData}
\usage{
generateNoiseRawData(cell_lines, drugs, save = TRUE)
}
\value{
data.frame with raw input data or MAE with processed data
}
\description{
generateNoiseRawData
}
\examples{

cell_lines <- create_synthetic_cell_lines()
drugs <- create_synthetic_drugs()
generateNoiseRawData(cell_lines, drugs)

}
\keyword{internal}
