% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateTripleComboMatrix}
\alias{generateTripleComboMatrix}
\title{generateTripleComboMatrix}
\usage{
generateTripleComboMatrix(cell_lines, drugs, save = TRUE)
}
\value{
data.table with raw input data or MAE with processed data
}
\description{
generateTripleComboMatrix
}
\keyword{internal}
