% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.utils.R
\name{get_synthetic_data}
\alias{get_synthetic_data}
\title{Get synthetic data from gDRtestData package}
\usage{
get_synthetic_data(rds)
}
\arguments{
\item{rds}{RDS filename}
}
\value{
loaded data
}
\description{
Get synthetic data from gDRtestData package
}
\examples{
data <- "finalMAE_small.RDS"
original <- get_synthetic_data(data)

}
