% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type.R
\name{identify_data_type}
\alias{identify_data_type}
\title{Identify type of data}
\usage{
identify_data_type(df, codilution_conc = 2, matrix_conc = 1)
}
\arguments{
\item{df}{data.frame of raw drug response data containing both treated and 
untreated values}

\item{codilution_conc}{integer of maximum number of concentration ratio of 
co-treatment to classify as codilution data type. Defaults to \code{2}.}

\item{matrix_conc}{integer of minimum number of concentration pairs of 
co-treatment to classify as co-treatment or matrix data type. Forces any 
co-treatment as a matrix data model. Defaults to \code{1}.}
}
\value{
data.frame of raw drug response data with additional column `type` 
with the info of data type for a given row of data.frame
}
\description{
Identify type of data
}
\author{
Bartosz Czech <bartosz.czech@contractors.roche.com>
}
