% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type.R
\name{identify_data_type}
\alias{identify_data_type}
\title{Identify type of data}
\usage{
identify_data_type(
  df,
  cotreatment_conc = 4,
  codilution_conc = 2,
  matrix_conc = 4
)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Identify type of data
}
