% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df.R
\name{map_untreated}
\alias{map_untreated}
\title{Identify untreated rows based on Drug treatment alone}
\usage{
map_untreated(mat_elem)
}
\arguments{
\item{mat_elem}{input data frame}
}
\value{
list
}
\description{
Identify untreated rows based on Drug treatment alone
}
\details{
Using the given rownames, map the untreated conditions
}
