% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{merge_data}
\usage{
merge_data(manifest, treatments, data)
}
\arguments{
\item{manifest}{a data frame with a manifest info}

\item{treatments}{a data frame with a treaatments info}

\item{data}{a data frame with a raw data info}
}
\value{
a dataframe with merged data and metadata.
}
\description{
Merge all the input data into a single data.frame
}
\examples{
dataDir <- system.file("extdata", "data1", package = "gDRimport")
manifest <- list.files(dataDir, pattern = "manifest", full.names = TRUE)
template <- list.files(dataDir, pattern = "Template", full.names = TRUE)
raw_data <- list.files(dataDir, pattern = "^RawData", full.names = TRUE)
l_data <- gDRimport::load_data(manifest, template, raw_data)
gDRcore::merge_data(
  l_data$manifest, 
  l_data$treatments, 
  l_data$data
)

}
