% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{merge_data}
\usage{
merge_data(manifest, treatments, data, collapse_Drugs = TRUE)
}
\arguments{
\item{manifest}{a data frame with a manifest info}

\item{treatments}{a data frame with a treaatments info}

\item{data}{a data frame with a raw data info}

\item{collapse_Drugs}{mode drug_2 to Drug when Drug == vehicle (default is TRUE)}
}
\value{
a dataframe with merged data and metadata.
}
\description{
Merge all the input data into a single data.frame
}
