% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{merge_data}
\usage{
merge_data(manifest, treatments, data)
}
\arguments{
\item{manifest}{a data frame with a manifest info}

\item{treatments}{a data frame with a treaatments info}

\item{data}{a data frame with a raw data info}
}
\value{
a dataframe with merged data and metadata.
}
\description{
Merge all the input data into a single data.frame
}
\examples{
td <- gDRimport::get_test_data()
l_tbl <- gDRimport::load_data(gDRimport::manifest_path(td), gDRimport::template_path(td), gDRimport::result_path(td))
merge_data(
  l_tbl$manifest, 
  l_tbl$treatments, 
  l_tbl$data
)

}
