% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_result_df}
\alias{order_result_df}
\title{Order_result_df}
\usage{
order_result_df(df_)
}
\arguments{
\item{df_}{a data.table with results}
}
\value{
a ordered data.table with results
}
\description{
Order a data.table with results
}
\keyword{utils}
