% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDrugResponseProcessingPipeline.R
\name{prepare_input.MultiAssayExperiment}
\alias{prepare_input.MultiAssayExperiment}
\title{Prepare input data common for all experiments}
\usage{
\method{prepare_input}{MultiAssayExperiment}(
  x,
  nested_confounders = gDRutils::get_SE_identifiers(x[[1]], "barcode"),
  nested_identifiers_l = .get_default_nested_identifiers(x[[1]]),
  raw_data_field = "experiment_raw_data"
)
}
\arguments{
\item{x}{MAE object with dose-reponse data}
}
\description{
Current steps
- refining nested confounders
- refining nested identifiers
- splitting df_ into (per experiment) df_list
}
