% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input.R
\name{prepare_input}
\alias{prepare_input}
\title{Prepare input data common for all experiments}
\usage{
prepare_input(x, ...)
}
\arguments{
\item{x}{data.frame with raw data or MAE object with dose-reponse data}

\item{...}{additional parameters}
}
\description{
Current steps
- refining nested confounders
- refining nested identifiers
- splitting df_ into (per experiment) df_list
}
