% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{apply_fit}
\alias{apply_fit_to_se}
\alias{apply_fits}
\alias{fit_drug_response_metrics}
\alias{fit_drug_response_metrics_4p}
\alias{get_fit_profile}
\alias{get_fit_profiles}
\alias{register_fit_profile}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{gDRutils}{\code{\link[gDRutils]{apply_fit}}, \code{\link[gDRutils]{apply_fit_to_se}}, \code{\link[gDRutils]{apply_fits}}, \code{\link[gDRutils]{fit_drug_response_metrics}}, \code{\link[gDRutils]{fit_drug_response_metrics_4p}}, \code{\link[gDRutils]{get_fit_profile}}, \code{\link[gDRutils]{get_fit_profiles}}, \code{\link[gDRutils]{register_fit_profile}}}
}}

