% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type.R
\name{split_raw_data}
\alias{split_raw_data}
\title{Split raw data into list based on the data types}
\usage{
split_raw_data(df, type_col = "type")
}
\arguments{
\item{df}{data frame with raw data}

\item{type_col}{column with identifier of data types in df}
}
\value{

}
\description{
Split raw data into list based on the data types
}
