% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type.R
\name{split_raw_data}
\alias{split_raw_data}
\title{Split raw data into list based on the data types}
\usage{
split_raw_data(df, type_col = "type")
}
\arguments{
\item{df}{data.frame of raw drug response data containing both treated and 
untreated values with column specified in `type_col` argument.}

\item{type_col}{string with column names in `df` with info about data type.
Defaults to \code{"type"}.}
}
\value{
list with split data based on its data type
}
\description{
Split raw data into list based on the data types
}
\author{
Bartosz Czech <bartosz.czech@contractors.roche.com>
}
