% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type.R
\name{split_raw_data}
\alias{split_raw_data}
\title{Split raw data into list based on the data types}
\usage{
split_raw_data(df, type_col = "type")
}
\arguments{
\item{df}{data.frame of raw drug response data containing both treated and untreated values with
column specified in `type_col` argument.}

\item{type_col}{string with column names in `df` with info about data type.
Defaults to \code{"type"}.}
}
\value{
list with split data based on its data type
}
\description{
Split raw data into list based on the data types
}
\examples{

#'  conc <- rep(seq(0, 0.3, 0.1), 2)
 ctrl_df <- S4Vectors::DataFrame(ReadoutValue = c(2, 2, 1, 1, 2, 1),
                               Concentration = rep(0, 6),
                               masked = FALSE,
                               DrugName = rep(c("DRUG_10", "vehicle", "DRUG_8"), 2),
                               CellLineName = "CELL1")

  trt_df <- S4Vectors::DataFrame(ReadoutValue = rep(seq(1, 4, 1), 2),
                              Concentration = conc,
                              masked = rep(FALSE, 8),
                              DrugName = c("DRUG_10", "DRUG_8"),
                              CellLineName = "CELL1")
  input_df <- as.data.frame(rbind(ctrl_df, trt_df))
  input_df$Duration <- 72
  input_df$CorrectedReadout2 <- input_df$ReadoutValue 
  split_df <- identify_data_type(input_df)
  split_raw_data(split_df)

}
\author{
Bartosz Czech <bartosz.czech@contractors.roche.com>
}
