% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.utils.R
\name{test_synthetic_data}
\alias{test_synthetic_data}
\title{Testing synthetic data form gDRtestData package}
\usage{
test_synthetic_data(
  original,
  data,
  dataName,
  override_untrt_controls = NULL,
  tolerance = 0.001
)
}
\arguments{
\item{original}{original MAE assay}

\item{data}{datase MAE or data frame}

\item{dataName}{dataset name}

\item{override_untrt_controls}{named list containing defining factors in 
the treatments}

\item{tolerance}{tolerance factor}
}
\value{
`NULL`
}
\description{
Testing synthetic data form gDRtestData package
}
\examples{
data <- "finalMAE_small.RDS"
original <- get_synthetic_data(data)
mae <- purrr::quietly(gDRtestData::generateNoiseRawData)(
  cell_lines, drugs, e_inf, ec50, hill_coef, FALSE
)
test_synthetic_data(original, mae$result, data)

}
